requirements.txt
tensorflow==2.1.0
tensorflow-probability==0.8.0
numpy==1.17.2
streamlit==0.57.3
Pillow==6.2.0
pyyaml==5.1.2
scipy==1.4.1
matplotlib==3.1.1
Pillow==6.2.0
